Extract extensions
