+++ source/dotnet/wholesale-api/WebApi/Extensions/DependencyInjection/ApiVersioningExtensions.cs
@@ -1,0 +1,49 @@
+﻿// Copyright 2020 Energinet DataHub A/S
+//
+// Licensed under the Apache License, Version 2.0 (the "License2");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using System.Runtime.CompilerServices;
+using Asp.Versioning;
+
+namespace Energinet.DataHub.Wholesale.WebApi.Extensions.DependencyInjection;
+
+/// <summary>
+/// Extension methods for <see cref="IServiceCollection"/>
+/// that allow adding API versioning services to an ASP.NET Core app.
+/// </summary>
+public static class ApiVersioningExtensions
+{
+    /// <summary>
+    /// Register services for enabling API versioning that defaults to the given version.
+    /// </summary>
+    public static IServiceCollection AddApiVersioningForWebApplication(this IServiceCollection services, ApiVersion defaultVersion)
+    {
+        services
+            .AddApiVersioning(options =>
+            {
+                // If client doesn't specify version, we assume the following default
+                options.DefaultApiVersion = defaultVersion;
+                options.AssumeDefaultVersionWhenUnspecified = true;
+                options.ReportApiVersions = true;
+            })
+            .AddApiExplorer(options =>
+            {
+                // API version format strings: https://github.com/dotnet/aspnet-api-versioning/wiki/Version-Format#custom-api-version-format-strings
+                options.GroupNameFormat = "'v'VVV";
+                options.SubstituteApiVersionInUrl = true;
+            });
+
+        return services;
+    }
+
+}
